+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/provider/foundationdb/indexes/StandardIndexMaintainer.java
@@ -28,6 +28,7 @@ import com.apple.foundationdb.async.AsyncUtil;
 import com.apple.foundationdb.async.MoreAsyncUtil;
 import com.apple.foundationdb.async.RangeSet;
 import com.apple.foundationdb.record.EvaluationContext;
+import com.apple.foundationdb.record.ExecuteProperties;
 import com.apple.foundationdb.record.IndexEntry;
 import com.apple.foundationdb.record.IndexScanType;
 import com.apple.foundationdb.record.IsolationLevel;
@@ -292,10 +293,9 @@ public abstract class StandardIndexMaintainer extends IndexMaintainer {
     protected <M extends Message> CompletableFuture<Void> updateIndexKeys(@Nonnull final FDBIndexableRecord<M> savedRecord,
                                                                           final boolean remove,
                                                                           @Nonnull final List<IndexEntry> indexEntries) {
-        for (IndexEntry entry : indexEntries) {
-            updateOneKey(savedRecord, remove, entry);
-        }
-        return AsyncUtil.DONE;
+        return CompletableFuture.allOf(indexEntries.stream()
+                .map(entry -> updateOneKeyAsync(savedRecord, remove, entry))
+                .toArray(CompletableFuture[]::new));
     }
 
     /**
@@ -304,10 +304,11 @@ public abstract class StandardIndexMaintainer extends IndexMaintainer {
      * @param savedRecord the record being indexed
      * @param remove <code>true</code> if removing from index
      * @param indexEntry the entry for the index to be updated
+     * @return a future completed when the key is updated
      */
-    protected <M extends Message> void updateOneKey(@Nonnull final FDBIndexableRecord<M> savedRecord,
-                                                    final boolean remove,
-                                                    @Nonnull final IndexEntry indexEntry) {
+    protected <M extends Message> CompletableFuture<Void> updateOneKeyAsync(@Nonnull final FDBIndexableRecord<M> savedRecord,
+                                                                            final boolean remove,
+                                                                            @Nonnull final IndexEntry indexEntry) {
         final Tuple valueKey = indexEntry.getKey();
         final Tuple value = indexEntry.getValue();
         final long startTime = System.nanoTime();
@@ -316,13 +317,15 @@ public abstract class StandardIndexMaintainer extends IndexMaintainer {
         final byte[] valueBytes = value.pack();
         if (remove) {
             state.transaction.clear(keyBytes);
-            if (state.store.isIndexWriteOnly(state.index) && state.index.isUnique()) {
-                updateUniquenessViolations(valueKey, savedRecord.getPrimaryKey(), null, true);
-            }
             if (state.store.getTimer() != null) {
                 state.store.getTimer().recordSinceNanoTime(FDBStoreTimer.Events.DELETE_INDEX_ENTRY, startTime);
                 state.store.countKeyValue(FDBStoreTimer.Counts.DELETE_INDEX_KEY, FDBStoreTimer.Counts.DELETE_INDEX_KEY_BYTES, FDBStoreTimer.Counts.DELETE_INDEX_VALUE_BYTES,
                         keyBytes, valueBytes);
+            }
+            if (state.store.isIndexWriteOnly(state.index) && state.index.isUnique()) {
+                return removeUniquenessViolationsAsync(valueKey, savedRecord.getPrimaryKey());
+            } else {
+                return AsyncUtil.DONE;
             }
         } else {
             checkKeyValueSizes(savedRecord, valueKey, value, keyBytes, valueBytes);
@@ -345,6 +348,7 @@ public abstract class StandardIndexMaintainer extends IndexMaintainer {
                 state.store.countKeyValue(FDBStoreTimer.Counts.SAVE_INDEX_KEY, FDBStoreTimer.Counts.SAVE_INDEX_KEY_BYTES, FDBStoreTimer.Counts.SAVE_INDEX_VALUE_BYTES,
                         keyBytes, valueBytes);
             }
+            return AsyncUtil.DONE;
         }
     }
 
@@ -357,6 +361,26 @@ public abstract class StandardIndexMaintainer extends IndexMaintainer {
         } else {
             state.transaction.set(uniquenessKeyBytes, (existingKey == null) ? new byte[0] : existingKey.pack());
         }
+    }
+
+    @Nonnull
+    @Override
+    public CompletableFuture<Void> removeUniquenessViolationsAsync(@Nonnull Tuple valueKey, @Nonnull Tuple primaryKey) {
+        Subspace uniqueValueSubspace = state.store.indexUniquenessViolationsSubspace(state.index).subspace(valueKey);
+        state.transaction.clear(uniqueValueSubspace.pack(primaryKey));
+        // Remove the last entry if it was the second last entry in the unique value subspace.
+        RecordCursor<KeyValue> uniquenessViolationEntries = KeyValueCursor.Builder.withSubspace(uniqueValueSubspace)
+                .setContext(state.context)
+                .setScanProperties(new ScanProperties(ExecuteProperties.newBuilder()
+                        .setReturnedRowLimit(2)
+                        .setIsolationLevel(IsolationLevel.SERIALIZABLE)
+                        .build()))
+                .build();
+        return uniquenessViolationEntries.getCount().thenAccept(count -> {
+            if (count == 1) {
+                state.transaction.clear(uniqueValueSubspace.pack());
+            }
+        });
     }
 
     @Override